Switched to merging context JIT at log-time.
Previously, calling withContext($context) would merge $context on top
of the parent's context and store it on the new instance. This caused
issues with monolog's withName() method, and generally confused things.

Now, context is not copied to descendent loggers, they simply track
whatever context was passed to withContext(), and maintain a reference
to their parent. At log-time, context is recursively merged into a
single array, and any callables are replaced with their results.

This avoids issues with shared context storage. One earlier attempt
to make withName() pay nicely was to copy the internal $context array
by reference. But that had unpleasant side-effects like causing context
added to a child to be persisted in the parent.

This may not be as performant as the previous implementation, but it's
actually correct. I hope.