Initial pass at just wrapping PSR3.
Back in PHP 7.4 days, when I started this library, I seem to recall this didn't
work out very well. But I can't recall precisely why, so I'm trying it again,
because it seems as though we can get much cleaner typing and static analysis,
and (and because of it being) a much simpler implementation.

While there may be some situations where it would be easier to integrate if your
StackLogger extended (instead of wrapping) you Logger, supporting that use-case
is probably not worth it. Outside of a factory, most places that currently call
for a Monolog\Logger are likely to be mostly only using PSR-3 methods. I will
need to think more about how people use withName() in monolog, though.